Sharing unstable work
Currently working on project and sharing unstable work. It contains a lot of error.
